split remote calls into train-interface.lua renamed surface-connections to surface-interface
