+++ changelog.txt
@@ -1,12 +1,14 @@
 ---------------------------------------------------------------------------------------------------
 Version: 1.18.0
-Date: 2022-10-13
+Date: 2022-10-27
   Features:
     - Create deliveries across surfaces, train surface transitions have to be handled by 3rd party mods
     - added remote call connect_surfaces
     - added remote call disconnect_surfaces
     - added remote call clear_all_surface_connections
     - added remote call reassign_delivery
+    - added remote call get_or_create_next_temp_stop
+    - added remote call get_next_logistic_stop
   Changes:
     - added on_dispatcher_updated.new_deliveries as replacement for on_delivery_created
     - removed event on_delivery_created, running in same tick as dispatcher produced microstutter
